% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bernoulli_M_step.R
\name{Bernoulli_M_step}
\alias{Bernoulli_M_step}
\title{Function to perform M step in EM algorithm.}
\usage{
Bernoulli_M_step(X, p_nk)
}
\arguments{
\item{X}{Matrix of binary data to cluster where in rows there are features and in columns observations.}

\item{p_nk}{Posterior probabilities of the hidden variables.}
}
\value{
Function returns a \code{dlist} which contains: \describe{
 \item{p}{Matrix of probabilites for each observation and cluster}
 \item{a}{Vector of each cluster weight in Bernoulli mixture model.}
}
}
\description{
Function description
}
