% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulliEM_ini.R
\name{bernoulliEM_ini}
\alias{bernoulliEM_ini}
\title{Function to perform initialization for EM step}
\usage{
bernoulliEM_ini(xdata, k, ini)
}
\arguments{
\item{xdata}{Matrix of binary data to cluster where in rows there are features and in columns observations.}

\item{k}{Number of clusters to divide your data.}

\item{ini}{Method for parameters initialization. There are three options 'random','kmeans' or "kmeanspp".}
}
\value{
Function returns a \code{dlist} which contains: \describe{
 \item{p}{Matrix of probabilites for each observation and cluster}
 \item{a}{Vecotr of each cluster weight in Bernoulli mixture model.}
}
}
\description{
Function description
}
\seealso{
\code{\link{Bernoulli_M_step}}
}
