% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{Exemplary binary data for clustering}
\format{
List
}
\source{
Randomly generated data
}
\usage{
data(example)
}
\description{
This data set was created ... bla bla bla
The variables are as follows:
}
\details{
\itemize{
  \item probs. Matrix of probabilities for each observation and cluster.
  \item alphas. Vector of each cluster weight in data.
  \item nouli_data. Binary data for clustering.
  \item gt. Ground truth of cluster assignment.
}
}
\keyword{dataset}
